test(bruno-cli): move quickjs trap fixtures to files behind a shared helper
The trap-containment collection was seven heredocs inside the spec. It now
lives under tests/integration/fixtures, copied into a temp dir by
createCollectionFixture, which substitutes the address the spec's server
bound to at runtime.
