+++ packages/bruno-cli/tests/integration/quickjs-trap-containment/fixtures/collection/04-async-test-callback.yml
@@ -1,0 +1,22 @@
+info:
+  name: 04-async-test-callback
+  type: http
+  seq: 4
+
+http:
+  method: GET
+  url: "{{baseUrl}}/first"
+
+runtime:
+  scripts:
+    - type: tests
+      code: |-
+        test('async test with awaited host work passes', async () => {
+          const val = await new Promise((resolve) => {
+            setTimeout(() => {
+              resolve('expected');
+            }, 120);
+          });
+          expect(val).to.equal('expected');
+          bru.setVar('asyncTestRan', 'yes');
+        });
