+++ packages/bruno-cli/tests/integration/quickjs-trap-containment/fixtures/collection/05-async-test-completed.yml
@@ -1,0 +1,16 @@
+info:
+  name: 05-async-test-completed
+  type: http
+  seq: 5
+
+http:
+  method: GET
+  url: "{{baseUrl}}/first"
+
+runtime:
+  scripts:
+    - type: tests
+      code: |-
+        test('the awaited test callback finished before its run returned', function() {
+          expect(bru.getVar('asyncTestRan')).to.equal('yes');
+        });
