+++ packages/bruno-cli/tests/integration/quickjs-trap-containment/fixtures/collection/bruno.json
@@ -1,0 +1,6 @@
+{
+  "version": "1",
+  "name": "quickjs-trap-containment",
+  "type": "collection",
+  "ignore": []
+}
