+++ packages/bruno-cli/tests/integration/quickjs-trap-containment/fixtures/collection/opencollection.yml
@@ -1,0 +1,3 @@
+opencollection: "1.0.0"
+info:
+  name: quickjs-trap-containment
